test: gate Wing internal response tests by platform
