+++ wing_feather_internal_test.go
@@ -1,0 +1,161 @@
+//go:build linux || darwin
+
+package kruda
+
+import (
+	"bytes"
+	"testing"
+)
+
+func TestWingResponsePlaintextModeUsesHandlerFastPath(t *testing.T) {
+	resp := acquireResponse()
+	defer releaseResponse(resp)
+
+	resp.responseMode = responsePlaintext
+	resp.SetStaticText(201, "text/plain; charset=utf-8", "created")
+
+	if resp.staticResp != nil {
+		t.Fatal("plaintext response mode should not use the shared static response cache")
+	}
+	if !resp.plaintextFast {
+		t.Fatal("plaintext response mode did not enable plaintext fast serialization")
+	}
+
+	data := resp.buildZeroCopy()
+	for _, want := range [][]byte{
+		[]byte("HTTP/1.1 201 Created\r\n"),
+		[]byte("Content-Type: text/plain; charset=utf-8\r\n"),
+		[]byte("Content-Length: 7\r\n"),
+		[]byte("\r\n\r\ncreated"),
+	} {
+		if !bytes.Contains(data, want) {
+			t.Fatalf("plaintext response missing %q in:\n%s", want, data)
+		}
+	}
+}
+
+func TestWingResponseGenericStaticTextKeepsStaticCache(t *testing.T) {
+	resp := acquireResponse()
+	defer releaseResponse(resp)
+
+	resp.SetStaticText(200, "text/plain; charset=utf-8", "ok")
+
+	if resp.staticResp == nil {
+		t.Fatal("generic Wing text path should keep using the shared static response cache")
+	}
+	if resp.plaintextFast {
+		t.Fatal("generic Wing text path should not enable plaintext response mode")
+	}
+}
+
+func TestWingPlaintextModeStillRunsHandlerMiddlewareLifecycle(t *testing.T) {
+	app := New(Wing())
+	var middlewareRan, beforeRan, handlerRan, afterRan bool
+	app.Use(func(c *Ctx) error {
+		middlewareRan = true
+		return c.Next()
+	})
+	app.BeforeHandle(func(c *Ctx) error {
+		beforeRan = true
+		return nil
+	})
+	app.AfterHandle(func(c *Ctx) error {
+		afterRan = true
+		return nil
+	})
+	app.Get("/plaintext", func(c *Ctx) error {
+		handlerRan = true
+		return c.Text("ok")
+	}, WingPlaintext())
+	app.Compile()
+
+	tr, ok := app.transport.(*Transport)
+	if !ok {
+		t.Skip("Wing transport unavailable on this platform")
+	}
+	f := tr.config.Feathers["GET /plaintext"]
+	if len(f.handlers) == 0 {
+		t.Fatal("WingPlaintext route did not retain its handler chain")
+	}
+
+	resp := acquireResponse()
+	defer releaseResponse(resp)
+	resp.responseMode = responsePlaintext
+
+	app.serveKrudaRoute(resp, &wingRequest{method: "GET", path: "/plaintext", keepAlive: true}, f.handlers)
+
+	if !middlewareRan || !beforeRan || !handlerRan || !afterRan {
+		t.Fatalf("middleware=%v before=%v handler=%v after=%v", middlewareRan, beforeRan, handlerRan, afterRan)
+	}
+	if !resp.plaintextFast {
+		t.Fatal("simple WingPlaintext handler did not use plaintext response mode")
+	}
+}
+
+func TestWingPlaintextModeGroupRouteRetainsHandlerChain(t *testing.T) {
+	app := New(Wing())
+	var groupMiddlewareRan, handlerRan bool
+	api := app.Group("/api").Use(func(c *Ctx) error {
+		groupMiddlewareRan = true
+		return c.Next()
+	})
+	api.Get("/plaintext", func(c *Ctx) error {
+		handlerRan = true
+		return c.Text("ok")
+	}, WingPlaintext())
+	app.Compile()
+
+	tr, ok := app.transport.(*Transport)
+	if !ok {
+		t.Skip("Wing transport unavailable on this platform")
+	}
+	f := tr.config.Feathers["GET /api/plaintext"]
+	if len(f.handlers) == 0 {
+		t.Fatal("WingPlaintext group route did not retain its handler chain")
+	}
+
+	resp := acquireResponse()
+	defer releaseResponse(resp)
+	resp.responseMode = responsePlaintext
+
+	app.serveKrudaRoute(resp, &wingRequest{method: "GET", path: "/api/plaintext", keepAlive: true}, f.handlers)
+
+	if !groupMiddlewareRan || !handlerRan {
+		t.Fatalf("groupMiddleware=%v handler=%v", groupMiddlewareRan, handlerRan)
+	}
+	if !resp.plaintextFast {
+		t.Fatal("simple grouped WingPlaintext handler did not use plaintext response mode")
+	}
+}
+
+func TestWingPlaintextModeCustomHeaderFallsBackToGenericResponse(t *testing.T) {
+	app := New(Wing())
+	app.Get("/plaintext", func(c *Ctx) error {
+		c.SetHeader("X-Test", "yes")
+		return c.Text("ok")
+	}, WingPlaintext())
+	app.Compile()
+
+	tr, ok := app.transport.(*Transport)
+	if !ok {
+		t.Skip("Wing transport unavailable on this platform")
+	}
+	f := tr.config.Feathers["GET /plaintext"]
+	if len(f.handlers) == 0 {
+		t.Fatal("WingPlaintext route did not retain its handler chain")
+	}
+
+	resp := acquireResponse()
+	defer releaseResponse(resp)
+	resp.responseMode = responsePlaintext
+
+	app.serveKrudaRoute(resp, &wingRequest{method: "GET", path: "/plaintext", keepAlive: true}, f.handlers)
+
+	if resp.plaintextFast {
+		t.Fatal("custom response headers must fall back to generic serialization")
+	}
+	data := resp.buildZeroCopy()
+	if !bytes.Contains(data, []byte("X-Test: yes\r\n")) {
+		t.Fatalf("generic fallback missing custom header:\n%s", data)
+	}
+}
